+++ cli/dts/lib.deno.window.d.ts
@@ -62,7 +62,7 @@ declare function confirm(message?: string): boolean;
  */
 declare function prompt(message?: string, defaultValue?: string): string | null;
 
-// TODO(nayeemrmn): Move this to `op_crates/web` where its implementation is.
+// TODO(nayeemrmn): Move this to `extensions/web` where its implementation is.
 // The types there must first be split into window, worker and global types.
 /** The location (URL) of the object it is linked to. Changes done on it are
  * reflected on the object it relates to. Accessible via
@@ -130,6 +130,6 @@ declare class Location {
   replace(url: string): void;
 }
 
-// TODO(nayeemrmn): Move this to `op_crates/web` where its implementation is.
+// TODO(nayeemrmn): Move this to `extensions/web` where its implementation is.
 // The types there must first be split into window, worker and global types.
 declare var location: Location;