+++ core/lib.rs
@@ -3,23 +3,25 @@ mod async_cancel;
 mod async_cell;
 mod bindings;
 pub mod error;
+mod extensions;
 mod flags;
 mod gotham_state;
 mod module_specifier;
 mod modules;
 mod normalize_path;
 mod ops;
+mod ops_builtin;
 mod ops_json;
 pub mod plugin_api;
 mod resources;
 mod runtime;
-mod zero_copy_buf;
 
 // Re-exports
 pub use futures;
 pub use rusty_v8 as v8;
 pub use serde;
 pub use serde_json;
+pub use serde_v8::Buffer as ZeroCopyBuf;
 pub use url;
 
 pub use crate::async_cancel::CancelFuture;
@@ -52,18 +54,18 @@ pub use crate::modules::ModuleSourceFuture;
 pub use crate::modules::NoopModuleLoader;
 pub use crate::modules::RecursiveModuleLoad;
 pub use crate::normalize_path::normalize_path;
-pub use crate::ops::op_close;
-pub use crate::ops::op_resources;
 pub use crate::ops::serialize_op_result;
 pub use crate::ops::Op;
 pub use crate::ops::OpAsyncFuture;
 pub use crate::ops::OpFn;
 pub use crate::ops::OpId;
 pub use crate::ops::OpPayload;
-pub use crate::ops::OpResponse;
+pub use crate::ops::OpResult;
 pub use crate::ops::OpState;
 pub use crate::ops::OpTable;
 pub use crate::ops::PromiseId;
+pub use crate::ops_builtin::op_close;
+pub use crate::ops_builtin::op_resources;
 pub use crate::ops_json::op_async;
 pub use crate::ops_json::op_sync;
 pub use crate::resources::Resource;
@@ -74,7 +76,9 @@ pub use crate::runtime::JsErrorCreateFn;
 pub use crate::runtime::JsRuntime;
 pub use crate::runtime::RuntimeOptions;
 pub use crate::runtime::Snapshot;
-pub use crate::zero_copy_buf::ZeroCopyBuf;
+// pub use crate::runtime_modules::include_js_files!;
+pub use crate::extensions::Extension;
+pub use crate::extensions::OpMiddlewareFn;
 
 pub fn v8_version() -> &'static str {
   v8::V8::get_version()